Move back validation errorsr into apartment type input
